# example_api_gateway/api/docs/swagger.yaml
definitions:
...
  models.Attribute:
    properties:
      attribute_types:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.CreatePositionRequest:
    properties:
      company_id:
        type: string
      name:
        type: string
      position_attributes:
        items:
          $ref: '#/definitions/models.PositionAttribute'
        type: array
      profession_id:
        type: string
    type: object
...
  models.GetAllPositionResponse:
    properties:
      count:
        type: integer
      positions:
        items:
          $ref: '#/definitions/models.Position'
        type: array
    type: object
...
  models.GetPositionAttribute:
    properties:
      attribute:
        $ref: '#/definitions/models.Attribute'
      attribute_id:
        type: string
      id:
        type: string
      position_id:
        type: string
      value:
        type: string
    type: object
  models.Position:
    properties:
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
      position_attributes:
        items:
          $ref: '#/definitions/models.GetPositionAttribute'
        type: array
      profession_id:
        type: string
    type: object
  models.PositionAttribute:
    properties:
      attribute_id:
        type: string
      value:
        type: string
    type: object
  models.PositionAttribute2:
    properties:
      id:
        type: string
      value:
        type: string
    type: object
...
  models.ResponseModel:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
  models.UpdatePosition:
    properties:
      resp:
        type: string
    type: object
  models.UpdatePositionRequest:
    properties:
      attribute:
        items:
          $ref: '#/definitions/models.Attribute'
        type: array
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
      position_attribute:
        items:
          $ref: '#/definitions/models.PositionAttribute2'
        type: array
      profession_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/position:
    get:
      consumes:
      - application/json
      description: Get All Position
      operationId: get-all-position
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      - description: profession_id
        in: query
        name: profession_id
        type: string
      - description: company_id
        in: query
        name: company_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.GetAllPositionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get all position
      tags:
      - position
    post:
      consumes:
      - application/json
      description: Create Position
      operationId: create-position
      parameters:
      - description: position
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/models.CreatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: create position
      tags:
      - position
  /v1/position/{position_id}:
    get:
      consumes:
      - application/json
      description: Get Position
      operationId: get-position
      parameters:
      - description: position_id
        in: path
        name: position_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Position'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get position
      tags:
      - position
  /v1/position/delete/{position_id}:
    delete:
      consumes:
      - application/json
      description: Delete Position
      operationId: delete-position
      parameters:
      - description: position_id
        in: path
        name: position_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.UpdatePosition'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: delete position
      tags:
      - position
  /v1/position/update/:
    put:
      consumes:
      - application/json
      description: Update Position
      operationId: update-position
      parameters:
      - description: position
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.UpdatePosition'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: update position
      tags:
      - position
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"